Add created_by validation

+++ todo/todo.py
@@ -19,7 +19,9 @@ bp = Blueprint('todo', __name__)
 def index(): 
     db, c = get_db()
     c.execute(
-        'select t.id, t.description, u.username, t.completed, t.created_at from todo t join user u on u.id = t.created_by order by created_at desc'
+        'select t.id, t.description, u.username, t.completed, t.created_at from todo t join user u on u.id = t.created_by '
+        'where t.created_by=%s order by created_at desc', 
+        (g.user['id'],)
     )
     todos = c.fetchall()
     return render_template('todo/index.html', todos=todos)
@@ -84,8 +86,8 @@ def update(id):
 
             c.execute( 
                 'update todo set description=%s, completed=%s'
-                ' where id=%s',
-                (description, completed, id)
+                ' where id=%s and created_by=%s',
+                (description, completed, id, g.user['id'])
             )
             db.commit() 
             return redirect(url_for('todo.index'))
@@ -96,6 +98,6 @@ def update(id):
 @login_required 
 def delete(id):
     db, c = get_db() 
-    c.execute('delete from todo where id=%s', (id,))
+    c.execute('delete from todo where id=%s and created_by=%s', (id, g.user['id']))
     db.commit() 
     return redirect(url_for('todo.index'))
